refactor: change place of db functions
